linker: never map pages as writable unless we really need to write
If we map a page as writable under a process that has the systemd
MemoryDenyWriteExecute property set to true, we can never make it
executable again. While this is not used much in practice, pulseaudio
does use it, and pulseaudio does not need any patching. So this patch
should, in theory, allow pulseaudio to start up without compromising
the security benefits from MemoryDenyWriteExecute.