[wip] egl/wayland_window: don't check buffer format/usage when reallocating
Something in this check seems to be wrong, because the buffer gets reallocated
constantly even though printing format and usage reveals them to be the same.

Since all buffers are created with the same format and usage anyway, this is
kinda redundant. But I am marking this as a WIP because we should rather find
out WHY the comparison is failing instead of just getting rid of it.